webapi/tct-contact-tizen-tests/contact/Contact_ContactEmailAddress_emails_types.html

<!DOCTYPE html>
<!--
Copyright (c) 2012 Intel Corporation. All rights reserved.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

*Redistributions of works must retain the original copyright notice, this list
of conditions and the following disclaimer.
*Redistributions in binary form must reproduce the original copyright notice,
this list of conditions and the following disclaimer in the documentation
and/or other materials provided with the distribution.
*Neither the name of Intel Corporation nor the names of its contributors
may be used to endorse or promote products derived from this work without
specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Cici Li <cici.x.li@intel.com>

Revision history:

-->

<html>
<head>
<title>Contact_ContactEmailAddress_emails_types</title>
<script type="text/javascript" src="../resources/testharness.js"></script>
<script type="text/javascript" src="../resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<script type="text/javascript">
//==== TEST: Contact_ContactEmailAddress_emails_types
//==== LABEL Check if can change ContactEmailAddress attribute: types
//==== PRIORITY P2
//==== SPEC Tizen Web API:Social:Contact:ContactEmailAddress:types A
//==== SPEC_URL https://developer.tizen.org/help/topic/org.tizen.web.device.apireference/tizen/contact.html
//==== TEST_CRITERIA
test(function () {
    var email = new tizen.ContactEmailAddress("user@domain.com", ["WORK"]);
    email.types[0] = "HOME";
    assert_not_equals(email, null, "email shouldn't be null");
    assert_equals(email.types[0], "HOME", "email.types[0] should be HOME");
}, document.title);

</script>
</body>
</html>
